Use absolute path to configure script instead of relative
The documentation of std::process::Command::current_dir() states that "If the
program path is relative [...] it's ambiguous whether it should be
interpreted relative to the parent's working directory or relative to
current_dir. The behavior in this case is platform specific and unstable".

Thus use absolute path to make sure the correct script is found and
gets executed.