Use MaybeUninit to manage uninitialized memory in pcap_compile
We need to allocate memory for libpcap::bpf_program, but allocation
leaves it in invalid state. We can wrap bpf_program in MaybeUninit
until the contents are initialized by pcap_compile(). After we are
sure pcap_compile() compiled the program we take the bpf_program out
from MaybeUninit and can be certain it's properly defined.