Add {Source, Destination}::into_inner()
The naming of unwrap() method for Source and Destination is a bit unfortunate
as unwrap() is almost like a reserved word. This leads into confusion in,
for example, code reviews where unwrap() might be confused with a call which
might lead to panic.

Thus, add into_inner() which does the same thing as unwrap() and is in line
with the way this kind of methods are typically named.

Also, deprecate the unwrap() allowing us to remove it in the future.